Don't override existing .gitconfig.user in setup script

+++ setup_git.sh
@@ -12,15 +12,19 @@ else
     echo Step skipped.
 fi
 
+if [ -f ~/.gitconfig.user ]; then
+    echo '~/.gitconfig.user' already exists.
+    echo Step skipped.
+else
+    echo Setup the .gitconfig.user '(press Enter to skip)'
+    echo Enter your name:
+    read name
+    if [ -n "${name}" ]; then    
+        echo Enter your email adress:
+        read email
 
-echo Setup the .gitconfig.user '(press Enter to skip)'
-echo Enter your name:
-read name
-if [ -n "${name}" ]; then    
-    echo Enter your email adress:
-    read email
-
-    echo "[user]" > ~/.gitconfig.user
-    echo -e '\t'name = $name >> ~/.gitconfig.user
-    echo -e '\t'email = $email >> ~/.gitconfig.user
+        echo "[user]" > ~/.gitconfig.user
+        echo -e '\t'name = $name >> ~/.gitconfig.user
+        echo -e '\t'email = $email >> ~/.gitconfig.user
+    fi
 fi
